pyproject.toml
[tool.poetry]
name = "open-notebook"
version = "0.0.1"
description = "An open source implementation of a research assistant, inspired by Google Notebook LM"
authors = ["Luis Novo <lfnovo@gmail.com>"]
license = "MIT"
readme = "README.md"
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
]

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.39.0"
watchdog = "^5.0.3"
pydantic = "^2.9.2"
loguru = "^0.7.2"
icecream = "^2.1.3"
langchain = "^0.3.3"
langgraph = "^0.2.38"
humanize = "^4.11.0"
streamlit-tags = "^1.2.8"
streamlit-scrollable-textbox = "^0.0.3"
tiktoken = "^0.8.0"
streamlit-monaco = "^0.1.3"
langgraph-checkpoint-sqlite = "^2.0.0"
pymupdf = "1.24.11"
python-magic = "^0.4.27"
langdetect = "^1.0.9"
youtube-transcript-api = "^0.6.2"
surrealdb = "^0.3.2"
openai = "^1.52.0"
pre-commit = "^4.0.1"
langchain-community = "^0.3.3"
langchain-openai = "^0.2.3"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
ruff = "^0.5.5"
mypy = "^1.11.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.isort]
profile = "black"
line_length = 88

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]

